LayoutTests/media/video-src-remove-expected.txt


EVENT(loadedmetadata)
EXPECTED (relativeURL(video.currentSrc) == 'content/test.mp4') OK
EXPECTED (isNaN(video.duration) == 'false') OK

RUN(video.removeAttribute('src'))

EXPECTED (relativeURL(video.currentSrc) == 'content/test.mp4') OK
EXPECTED (isNaN(video.duration) == 'false') OK

END OF TEST
Test that removing valid 'src' attribute DOES NOT trigger load of <source> elements
